Update instagram_clone_coding practice

+++ Django_study/instagram_clone_coding/posts/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import render, redirect
 from .models import Post
 from .forms import PostForm
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_http_methods, require_POST, require_safe
+
 
 # Create your views here.
+@require_safe
 def index(request):
     posts = Post.objects.all()
     context = {
@@ -11,6 +15,8 @@ def index(request):
     return render(request, 'posts/index.html', context)
 
 
+@login_required
+@require_http_methods(['GET', 'POST'])
 def create(request):
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES)
@@ -23,14 +29,19 @@ def create(request):
         'form': form
     }
     return render(request, 'posts/create.html', context)
-    
 
+
+@require_POST
 def delete(request, pk):
-    post = Post.objects.get(pk=pk)
-    post.delete()
-    return redirect('posts:index')
+    if request.user.is_authenticated:
+        post = Post.objects.get(pk=pk)
+        post.delete()
+        return redirect('posts:index')
+    return redirect('accounts:login')
 
 
+@login_required
+@require_http_methods(['GET', 'POST'])
 def update(request, pk):
     post = Post.objects.get(pk=pk)
     if request.method == 'POST':
